Move reused log retention components to own folder
- e.g., logic file, helpers used by other views & not just settings
