+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/log_retention/index.ts
@@ -1,0 +1,9 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+export { LogRetentionLogic } from './log_retention_logic';
+export * from './types';
+export * from './messaging';
