+++ x-pack/plugins/enterprise_search/public/applications/app_search/components/log_retention/messaging/determine_tooltip_content.ts
